small bug fix for the camera model

+++ include/camera.h
@@ -48,7 +48,7 @@ public:
 
     int countVisible(float x_w, float y_w, float theta_rad_w) const;
 
-    bool IsStateVisiblilty(float x_w, float y_w, float theta_rad_w);
+    bool IsStateVisiblilty(double x_w, double y_w, double theta_rad_w);
 
     void update_map(std::vector<ORB_SLAM2::MapPoint*> &vpPts);
